+++ CHANGELOG.md
@@ -20,7 +20,7 @@ Note that this project **does not** adhere to [Semantic Versioning](http://semve
 - An SLR can now be started from the SLR itself. [#9131](https://github.com/JabRef/jabref/pull/9131), [koppor#601](https://github.com/koppor/jabref/issues/601)
 - Implement a new ISBN Fetcher ([doi-to-bibtex-converter.herokuapp.com](http://doi-to-bibtex-converter.herokuapp.com) as source). [#9145](https://github.com/JabRef/jabref/pull/9145)
 - We added support for the Ukrainian and Arabic languages. [#9236](https://github.com/JabRef/jabref/pull/9236), [#9243](https://github.com/JabRef/jabref/pull/9243)
-- Changing a library's general file directory will be reflected in the default directory for unlinked file lookup, [koppor#546](https://github.com/koppor/jabref/issues/546)
+- Changing the default file directory will be reflected in the default directory for unlinked file lookup. [koppor#546](https://github.com/koppor/jabref/issues/546)
 
 ### Changed